+++ tunnel.c
@@ -138,7 +138,8 @@ static void on_xserver_read(uv_stream_t* stream, ssize_t nread, const uv_buf_t* 
          *'on_tclient_write' callback will do it.
          */
     } else if (nread < 0) {
-        xlog_debug("disconnected from proxy server: %s.", uv_err_name(nread));
+        xlog_debug("disconnected from proxy server: %s.",
+            uv_err_name((int) nread));
 
         uv_close((uv_handle_t*) stream, on_io_closed);
         uv_close((uv_handle_t*) &ctx->io_tclient, on_io_closed);
@@ -204,7 +205,7 @@ static void on_tclient_read(uv_stream_t* stream, ssize_t nread, const uv_buf_t* 
          *'on_xserver_write' callback will do it.
          */
     } else if (nread < 0) {
-        xlog_debug("disconnected from tunnel client: %s.", uv_err_name(nread));
+        xlog_debug("disconnected from tunnel client: %s.", uv_err_name((int) nread));
 
         uv_close((uv_handle_t*) stream, on_io_closed);
         uv_close((uv_handle_t*) &ctx->io_xserver, on_io_closed);
